+++ PokemonGo.RocketAPI.Logic/Shared/ISettings.cs
@@ -48,7 +48,7 @@ namespace PokeMaster.Logic.Shared
         int XPFarmedLimit { get; set; }
         int BreakInterval { get; set; }
         int BreakLength { get; set; }
-        int MinWalkSpeed { get; set; }
+        double MinWalkSpeed { get; set; }
         int PokemonCatchLimit { get; set; }
         int PokestopFarmLimit { get; set; }
         int navigation_option { get; set; }
